+++ tests/Feature/Concurrency/ConcurrentApiKeyUsageTest.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Str;
 beforeEach(function () {
     StudioApiRouteRegistrar::register();
 
-    $this->collection = StudioCollection::factory()->create(['slug' => 'posts']);
+    $this->collection = StudioCollection::factory()->apiEnabled()->create(['slug' => 'posts']);
 
     $this->titleField = StudioField::factory()->required()->create([
         'collection_id' => $this->collection->id,